refactor: split project-info page into smaller components
- Created project-info.ts for shared types
- Split ProjectSetupPanel into its own component
- Split ProjectDetailsPanel into its own component
- Split ChatPanel into its own component
- Split SaveLoadPanel into its own component
- Updated main page.tsx to use these components
